+++ tests/integration/test_objector.py
@@ -10,9 +10,8 @@ class TestObjector(IntegrationTest):
         message = 'USER_REQUIRED: \'Please log in to do that.\''
         with self.recorder.use_cassette(
                 'TestObjector.test_raise_api_exception'):
-            subreddit = self.reddit.subreddit(
-                pytest.placeholders.test_subreddit)
+            submission = self.reddit.submission('4b536h')
             with pytest.raises(APIException) as excinfo:
-                subreddit.mod.approve(self.reddit.submission('4b536h'))
+                submission.mod.approve()
             assert excinfo.value.error_type == 'USER_REQUIRED'
             assert str(excinfo.value) == message